@@ -1,13 +1,15 @@
 {
   "name": "eslint-plugin-ng-module-sort",
-  "version": "1.3.2",
+  "version": "1.4.0",
   "description": "Sort Angular and NestJS module imports, declarations, exports, controls, etc.",
   "keywords": [
     "eslint",
     "eslintplugin",
     "eslint-plugin",
     "angular",
-    "nestjs"
+    "nestjs",
+    "angular-sort",
+    "angular-import-sort"
   ],
   "author": "Daniel Lazar",
   "repository": {
@@ -37,22 +39,21 @@
     "postpack": "pinst --enable"
   },
   "dependencies": {
-    "requireindex": "^1.2.0"
+    "@typescript-eslint/utils": "8.48.1"
   },
   "devDependencies": {
-    "@types/eslint": "^8.37.0",
-    "@types/node": "^18.15.11",
-    "@typescript-eslint/parser": "^5.57.1",
-    "@typescript-eslint/types": "^5.57.1",
-    "@typescript-eslint/utils": "^5.57.1",
-    "eslint": "^8.19.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-doc-generator": "^1.0.0",
-    "eslint-plugin-eslint-plugin": "^5.0.0",
-    "eslint-plugin-node": "^11.1.0",
+    "@eslint/js": "^9.39.1",
+    "@types/eslint": "^9.6.1",
+    "@types/node": "^24.10.1",
+    "@typescript-eslint/parser": "8.48.1",
+    "@typescript-eslint/rule-tester": "8.48.1",
+    "eslint": "^9.39.1",
+    "eslint-doc-generator": "^2.4.0",
+    "eslint-plugin-eslint-plugin": "^7.2.0",
+    "eslint-plugin-n": "^17.23.1",
     "husky": "^9.0.11",
-    "mocha": "^10.0.0",
-    "nodemon": "^2.0.22",
+    "mocha": "^11.7.5",
+    "nodemon": "^3.1.11",
     "npm-run-all": "^4.1.5",
     "pinst": "^3.0.0",
     "prettier": "3.3.2",
@@ -60,10 +61,10 @@
     "typescript": "^5.0.4"
   },
   "engines": {
-    "node": "^14.17.0 || ^16.0.0 || >= 18.0.0"
+    "node": "^18.18.0 || ^20.9.0 || >=21.1.0"
   },
   "peerDependencies": {
-    "eslint": ">=7"
+    "eslint": ">=8.57.0"
   },
   "license": "MIT"
 }
