@@ -1,13 +1,13 @@
-import {Identifier} from '@typescript-eslint/types/dist/generated/ast-spec';
+import {TSESTree} from '@typescript-eslint/utils';
 
 import {DecoratorArrayItemsRuleContext} from '../../types';
 
 export const orderCheck = (
   context: DecoratorArrayItemsRuleContext,
-  elements: Array<Identifier>,
+  elements: Array<TSESTree.Identifier>,
   reverseSort: boolean,
 ): boolean => {
-  const sourceCode = context.getSourceCode();
+  const sourceCode = context.sourceCode;
   const elementNames = elements.map((el) => sourceCode.getText(el));
   const orderedElementNames = [...elementNames].sort();
   if (reverseSort) orderedElementNames.reverse();
